# Specialize Vec::extend to Vec::extend_from_slice