@fhartwig 2016-10-11
I'm not really sure what to do with the docstring on `extend_from_slice`. Saying that `extend` specialises to the same thing seems premature, considering that specialisation isn't stable yet. But leaving it probably also isn't ideal.