@apasel422 2016-10-12
Out of curiosity, what were the conflicts with `SpecExtend`?