@fhartwig 2016-10-12
@apasel422 As far as I remember, I tried to do
`impl <'a, T: 'a + Copy, I: IntoIterator<Item=&'a T>> SpecExtend<I> for Vec<T>`
and
`impl<'a, T: Copy> SpecExtend<&'a[T]> for Vec<T>`
and both of these would get a `conflicting implementation` error pointing to the `impl<I: IntoIterator> SpecExtend<I> for Vec<I::Item>` that comes before. I might be getting this wrong though, but I can look it up tonight.