@alexcrichton 2016-10-13
@bors: r+

Thanks!